ML_for_Battery_Design/src/simulation/simulation_model.py
from abc import ABC  # , abstractmethod


class SimulationModel(ABC):
    """Simulation model abstract class for hidden parameter sampling and simulation data generation

    Attributes:
        hidden_params (dict): boolean values for if each hidden parameter should be sampled or stay constant
        simulation_settings (dict): settings for generating simulation data
        sample_boundaries (dict): sampling boundaries for each hidden parameter
        default_param_values (dict): default values of hidden parameters, if not sampled
    """

    def __init__(
        self,
        hidden_params: dict,
        simulation_settings: dict,
        sample_boundaries: dict,
        default_param_values: dict,
    ) -> None:
        """Initializes parent :class:SimulationModel

        Args:
            hidden_params (dict): boolean values for if each hidden parameter should be sampled or stay constant
            simulation_settings (dict): settings for generating simulation data
            sample_boundaries (dict): sampling boundaries for each hidden parameter
            default_param_values (dict): default values of hidden parameters, if not sampled
        """
        self.hidden_params = hidden_params
        self.simulation_settings = simulation_settings
        self.sample_boundaries = sample_boundaries
        self.default_param_values = default_param_values
